Rename:AlertDialogComponent 위치 수정
